requirements.txt
gunicorn
flask
pandas
numpy
tensorflow
Pillow